fix: cycle refactor 및 팝콘작 불러오기
