+++ Domain/src/main/java/com/example/domains/diverseMovie/schedule/DiverseMovieSchedule.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 public class DiverseMovieSchedule {
     private final DiverseMovieService diverseMovieService;
      //매 주 월요일 0시 0분에 실행되도록 cron 설정
-     @Scheduled(cron = "0 0 1 * * WED")
+     @Scheduled(cron = "0 0 1 * * SUN")
      public void scheduledMethod() {
          // 수요일에 실행될 로직을 여기에 작성
          diverseMovieService.getTopRated();
